requirements.txt
fpdf2>=2.7
google-genai>=1.0.0
python-dotenv>=1.0.0
